Add initial support for NXP RT10XX
Surprise, the RT10xx USB peripheral is based on the LPC43xx one!
Except there's more endpoints and the EP0 control register is separated
from the rest in the struct.
Seems there is more to be done though, because while the hid example
works, the cdc_msc fails.
But it's a good start!